displaying message if it has modified files

+++ pypacking/pypacking.py
@@ -139,15 +139,18 @@ class PyPacking:
                     self.file_hashes[filepath] = new_hash
                     modified_files.append(filepath)
 
-            print(f'{len(modified_files)} files were changed. Copying...')
+            if modified_files:
+                print(f'{len(modified_files)} files were changed. Copying...')
 
-            for file in modified_files:
-                filepath = os.path.join('build', file)
-                print(f'Removing {file} from build/ directory...')
-                os.remove(filepath)
-                print(f'Copying {file} to build/{self.package_path} directory...', end='')
-                shutil.copyfile(file, filepath)
-                print('done')
+                for file in modified_files:
+                    filepath = os.path.join('build', file)
+                    print(f'Removing {file} from build/ directory...')
+                    os.remove(filepath)
+                    print(f'Copying {file} to build/{self.package_path} directory...', end='')
+                    shutil.copyfile(file, filepath)
+                    print('done')
+            else:
+                print('No files changed.')
         else:
             print('First build detected')
             print('Copying package to build/ directory...', end='')
